pyproject.toml
[tool.poetry]
name = "ml_cat_project"
version = "0.0.1"
description = "A small Machine Learning toy project to learn ML principles. It aims to recognize cats on a picture."
authors = ["Edouard Brülhart <edouard.brulhart@gmail.com>"]
repository = "https://github.com/edouardbruelhart/ml-cat-project"
documentation = "https://edouardbruelhart.github.io/ml-cat-project/"
readme = "README.md"
packages = [
  {include = "ml_cat_project"}
]

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
pillow = "^11.3.0"
torch = "^2.8.0"
torchvision = "^0.23.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
deptry = "^0.16.2"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
tox = "^4.11.1"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^9.2.7"
mkdocstrings = {extras = ["python"], version = "^0.26.1"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["ml_cat_project"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["ml_cat_project"]


[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
